refactor : types폴더 하나로 통일
